Update cptbc-frontend.php for using background-images to not set the size of the div to zero.
When using the CSS background images option the height option is used to set the height of the carousel images. Besides there are cases where you want to set the height to a particular value, 0 does not make sense at all. In this case every carousel item is set to "height:0px" and you do not see anything. The code change suggests an if statement which checks the value of the option 'Height if using bkgrnd images (px)' ($atts['background_images_height']) for not beeing 0. If it is zero, there is no ouput, otherwise it will echo the saved value.