Added video support for YouTube & Vimeo (#85)
* Set unique name for carousel categories

* Added fields to add video & aspect ratio.

Starting to add support for video embeds in the carousel. We need both the embed URL & the aspect ratio so that the height doesn't end up really badly set when the carousel hasn't had a fixed height set.

* Added in videos to carousel

Videos can now be added to the carousel - they replace the image & text content. Youtube & Vimeo videos support auto-pausing when the carousel slides away. Some cleanup could be done - I'll do this in the next set of changes.

* Minor cleanup of autopause JS

* Fixed whitespace

* Update cptbc-frontend.php